move selected filters into award amount component due to margin differences
