+++ src/js/containers/search/filters/awardAmount/AwardAmountSearchContainer.jsx
@@ -8,13 +8,9 @@ import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { is } from 'immutable';
-import { each } from 'lodash';
 
 import * as searchFilterActions from 'redux/actions/search/searchFilterActions';
 import AwardAmountSearch from 'components/search/filters/awardAmount/AwardAmountSearch';
-import SelectedAwardAmountBound from
-    'components/search/filters/awardAmount/SelectedAwardAmountBound';
-import { formatAwardAmountRange } from 'helpers/awardAmountHelper';
 
 
 const propTypes = {
@@ -52,37 +48,13 @@ export class AwardAmountSearchContainer extends React.Component {
         });
     }
 
-    stagedFilters() {
-        const filterObject = this.props.awardAmounts.toObject();
-        let stagedFilter;
-        let name;
-        each(filterObject, (val, key) => {
-            stagedFilter = val;
-            name = key;
-        });
-        if (!stagedFilter) return null;
-        const label = formatAwardAmountRange(stagedFilter);
-        return (
-            <SelectedAwardAmountBound
-                removeFilter={this.removeFilter}
-                name={name}
-                label={label} />
-        );
-    }
-
     render() {
-        const stagedFilters = this.stagedFilters();
         return (
             <div className="award-amount-filter">
                 <AwardAmountSearch
                     dirtyFilters={this.dirtyFilters()}
                     awardAmounts={this.props.awardAmounts}
                     selectAwardRange={this.selectAwardRange} />
-                <div
-                    className="selected-filters"
-                    role="status">
-                    {stagedFilters}
-                </div>
             </div>
         );
     }
